Added the checkHorizontal and checkVertical functions. These functions are called after every player's new action to check a four-in-a-row pattern (horizontally or vertically) was created.
checkHorizontally:
For each one of the 6 rows of the board checks if 4 neigbouring node-discs have
the same color (not black and not undefined). Returns true if yes.

checkVertically:
For each one of the 7 columns of the board checks if 4 neighbouring node-discs have
the same color (not black and not undefined). Returns true if yes.